Move InceptionResnet to components/cnn_convolutions.py
icecube_dnn.py now contains only the IceCubeDNN model itself; all
reusable convolution blocks live in components.
